virtio-msg: Add FF-A bus transport implementation
Introduce a virtio-msg bus implementation based on the Arm FF-A
(Firmware Framework for Arm) communication interface.

This bus enables virtio-msg transport over secure channels typically
used between the normal world OS and a secure OS or hypervisor. It
leverages the standardized FF-A interface to exchange messages with a
remote backend service, integrating with the core virtio-msg transport
and using FF-A service calls to transmit and receive messages.

Optionally, this bus supports attaching a reserved-memory region to
constrain DMA-coherent and streaming DMA allocations to a well-defined
contiguous area. This memory can be pre-mapped on the remote side,
reducing runtime overhead and avoiding accidental sharing of unrelated
pages due to page-granularity mapping.

To enable reserved memory, define a device tree node such as:

  reserved-memory {
    #address-cells = <2>;
    #size-cells   = <2>;
    ranges;

    rmem@100000000 {
      compatible = "restricted-dma-pool", "virtio-msg,ffa";
      reg = <0x00000001 0x00000000 0x0 0x00400000>; /* 4 MiB */
    };
  };

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>