firmware: arm_ffa: add public interfaces
Export some public interfaces to let an other driver:
- retrieve the FF-A version negociated
- setup/cleanup partitions
- setup/cleanup rxtx buffers

Make sure that all public interfaces properly fail if the FF-A driver
init did not succeed by checking drv_info not being NULL at the beginning
of each of them.

Also introduce an irq callback and a public function to register or
unregister a callback.
When a callback is registered it is called when a SRI or NRI is received
instead of going through the normal processing.

All those changes are introduced to create a driver usable to validate
the compliance to FF-A specification of an underlying hypervisor with
the FF-A driver functional when the tests are not executed.

Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>