firmware: arm_ffa: Add ACS for FF-A driver
Introduce a driver that can be used by the Architecture Compliance Suite
to validate the compliance to the FF-A specification of the system under
Linux (hypervisor, firmware, secure world).

The driver is providing a IOCTL user interface to:
- send/receive FF-A HVC/SMC calls
- allocate contiguous memory that can be later mmapped into the
  application
- wait for a SRI or NRI interrupt

When the interface is opened by a user application, the acs driver will
unload all FF-A driver by removing all detected partitions and will then
unregister the RXTX buffer. When the interface is closed, the partitions
are re-enumerated and the RXTX buffer is re-registered.

The driver can be built as a module.

Signed-off-by: Bertrand Marquis <bertrand.marquis@arm.com>