mailbox: pcc: Always clear the platform ack interrupt first
The PCC mailbox interrupt handler (pcc_mbox_irq()) currently checks
for command completion flags and any error status before clearing the
interrupt.

The below sequence highlights an issue in the handling of PCC mailbox
interrupts, specifically when dealing with doorbell notifications and
acknowledgment between the OSPM and the platform where type3 and type4
channels are sharing the interrupt.

        Platform Firmware                          OSPM/Linux PCC driver
--------------------------------------------------------------------------------
                                             build message in shmem
                                             ring type3 channel doorbell
receives the doorbell interrupt
  process the message from OSPM
  build response for the message
ring the platform ack interrupt to OSPM
					--->
start to build notification in type4 channel
                                             start processing in pcc_mbox_irq()
                                              enter pcc handler for type4 chan
                                                 command complete cleared
						 read the notification
                                        <---     clear platform ack irq
  * no effect from above as platform ack irq not yet triggered on this channel *
ring the platform ack interrupt on type4 channel
					--->
                                              leave pcc handler for type4 chan
                                              enter pcc handler for type3 chan
                                                 command complete set
						 read the response
                                        <---     clear platform ack irq
                                              leave pcc handler for type3 chan
                                             leave pcc_mbox_irq() handler
                                             start processing in pcc_mbox_irq()
                                              enter pcc handler for type4 chan
                                              leave pcc handler for type4 chan
                                              enter pcc handler for type3 chan
                                              leave pcc handler for type3 chan
                                             leave pcc_mbox_irq() handler

The key issue occurs when OSPM tries to acknowledge platform ack interrupt
for a notification which is ready to be read and processed but the interrupt
itself is not yet triggered by the platform.

This ineffective acknowledgment leads to an issue later in time where the
interrupt remains pending as we exit the interrupt handler without clearing
the platform ack interrupt as there is no pending response or notification.
The interrupt acknowledgment order is incorrect.

To resolve this issue, the platform acknowledgment interrupt should always
be cleared before processing the interrupt for any notifications or response.

Reported-by: Robbie King <robbiek@xsightlabs.com>
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>