mailbox: pcc: Always map the shared memory communication address
Currently the shared memory communication address was mapped by the
mailbox client drivers leading to all sorts of inconsistencies.

It also has resulted in the inconsistent attributes used while mapping
the shared memory regions.

In order to remove/eliminate any issues, let us ensures the shared
memory address is always mapped and unmapped when the PCC channels are
requested and release.

We need to map them as the ACPI PCCT associates these shared memory
with each channel subspace and may need use the headers in those
memory to manage the transport.

Since there are no users of pcc_chan_ioremap() and also it is mapped
by default, we can stop exporting it and make it static function.

Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>