cpufreq: Allow 'setpolicy' drivers to enable EAS
Some cpufreq drivers, like intel_pstate, have built-in governors that
are used instead of regular cpufreq governors, schedutil in particular,
but they can work with EAS just fine, so allow EAS to be used with
those drivers.

The underlying assumption is that if such a driver is used, EAS will
have to be enabled by it directly in which case the drivers can be
expected to know what it is doing.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>